+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Fixed
 - C++ query: missing `@container` capture in one pattern
 - C++ query: duplicate pattern for `initializer_list`
 - Common Lisp query: duplicate extmarks on some delimiters
+- Luadoc query: Remove broken pattern for dictionary key type
 - Ruby query: duplicate extmarks on some delimiters
 - Zig query: duplicate extmarks on some delimiters
 - Make a better effort to make parentheses in Vim script expressions work (there